hf/agent_generator_pkg/pyproject.toml
# ────────────────────────────────────────────────────────────────
#  pyproject.toml  •  build + metadata for agent-generator
# ────────────────────────────────────────────────────────────────

[build-system]
requires      = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name            = "agent-generator"
version         = "0.2.1"
description     = "Turn plain-English requirements into production-ready multi-agent AI projects (CrewAI, LangGraph, WatsonX Orchestrate)."
readme          = "README.md"
requires-python = ">=3.10"
license         = { file = "LICENSE" }
authors         = [
  { name = "Ruslan M. V.", email = "contact@ruslanmv.com" }
]
keywords = [
  "multi-agent",
  "crewai",
  "langgraph",
  "watsonx",
  "orchestrate",
  "LLM",
  "code-generation",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Framework :: FastAPI",
  "Topic :: Software Development :: Code Generators",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
]

# -----------------------------------------------------------------
# Core runtime dependencies
# -----------------------------------------------------------------
dependencies = [
  "pydantic>=2.12,<3",
  "pydantic-settings>=2.13,<3",
  "python-dotenv>=1.0",
  "typer>=0.24,<1",
  "jinja2>=3.1,<4",
  "requests>=2.33,<3",
  "rich>=14.0,<15",
  "structlog>=25.0,<26",
  "pyyaml>=6.0,<7",
]

# -----------------------------------------------------------------
# Optional extras
# -----------------------------------------------------------------
[project.optional-dependencies]
openai = [
  "openai>=2.30,<3",
  "tiktoken>=0.12,<1",
]
crewai = [
  "crewai>=1.12,<2",
]
langgraph = [
  "langgraph>=1.1,<2",
  "langchain-core>=0.1",
]
web = [
  "fastapi>=0.135,<1",
  "uvicorn>=0.42,<1",
  "python-multipart>=0.0.18",
  "requests>=2.33,<3",
]
all = [
  "agent-generator[openai,crewai,langgraph,web]",
]
release = [
  "build>=1.2",
  "twine>=5.0",
]
dev = [
  "pytest>=8.2",
  "pytest-xdist>=3.5",
  "pytest-asyncio>=0.24",
  "mypy>=1.10",
  "ruff>=0.4",
  "black>=24.4",
  "isort>=5.0",
  "pytest-cov>=5.0",
  "pre-commit>=3.7",
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
]

# -----------------------------------------------------------------
# CLI entry-point
# -----------------------------------------------------------------
[project.scripts]
agent-generator = "agent_generator.cli:_main"

# -----------------------------------------------------------------
# URLs
# -----------------------------------------------------------------
[project.urls]
Homepage   = "https://github.com/ruslanmv/agent-generator"
Repository = "https://github.com/ruslanmv/agent-generator"
Issues     = "https://github.com/ruslanmv/agent-generator/issues"

# -----------------------------------------------------------------
# Ruff configuration
# -----------------------------------------------------------------
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["F841", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["agent_generator"]

# -----------------------------------------------------------------
# Pytest configuration
# -----------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# -----------------------------------------------------------------
# MyPy configuration
# -----------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

# -----------------------------------------------------------------
# Hatchling configuration
# -----------------------------------------------------------------
[tool.hatch.build.targets.wheel]
packages = ["src/agent_generator"]
